bonus: darshan tests no longer fail if darshan-parser is absent

+++ tests/test_bin_cache_connector.py
@@ -7,12 +7,14 @@ methods.
 """
 
 import os
-import nose
+import errno
 import json
 import pandas
 import tempfile
 import StringIO
 import subprocess
+import nose
+import tokio.connectors.darshan
 
 def verify_json(json_str):
     data = json.loads(json_str)
@@ -93,31 +95,47 @@ CACHE_CONNECTOR_CONFIGS = [
 ]
 
 TEMP_FILE = None
+SKIP_DARSHAN = False # if darshan-parser isn't present
 FNULL = None
 
-def setup_tmpfile():
+def setup():
     global TEMP_FILE
+    global SKIP_DARSHAN
     global FNULL
     TEMP_FILE = tempfile.NamedTemporaryFile(delete=False)
     FNULL = open(os.devnull, 'w')
+    try:
+        output_str = subprocess.check_output(tokio.connectors.darshan.DARSHAN_PARSER_BIN)
+    except OSError as error:
+        if error[0] == errno.ENOENT:
+            SKIP_DARSHAN = True
+    except subprocess.CalledProcessError:
+        # this is ok--there's no way to make darshan-parser return zero without
+        # giving it a real darshan log
+        pass
 
-def teardown_tmpfile():
+def teardown():
     global TEMP_FILE
     global FNULL
-    TEMP_FILE.close()
-    os.unlink(TEMP_FILE.name)
+    if not TEMP_FILE.closed:
+        TEMP_FILE.close()
+    if os.path.exists(TEMP_FILE.name):
+        os.unlink(TEMP_FILE.name)
     FNULL.close()
 
 def run_cache_connector(binary, args, validators):
     global TEMP_FILE
 
+    if SKIP_DARSHAN and binary.endswith('cache_darshan.py'):
+        raise nose.SkipTest("%s not available for %s" % (
+                            tokio.connectors.darshan.DARSHAN_PARSER_BIN,
+                            binary))
+
     ### first test caching to stdout
     cmd = [ binary ] + args
     print "Executing:", cmd
     print "Caching to stdout..."
-    p = subprocess.Popen(cmd, stdout=subprocess.PIPE)
-    output_str = p.communicate()[0]
-    assert p.returncode == 0
+    output_str = subprocess.check_output(cmd)
     for validator in validators:
         validator(output_str)
 
@@ -130,7 +148,7 @@ def run_cache_connector(binary, args, validators):
     for validator in validators:
         validator(output_str)
 
-@nose.tools.with_setup(setup_tmpfile, teardown_tmpfile)
+@nose.tools.with_setup(setup, teardown)
 def test():
     """
     Test all connector cache scripts
